remove references to old Article (and Artefact)
