+++ server/model/article.js
@@ -1,43 +1,53 @@
-import entities from 'entities';
-import {createPerson, type Person} from './person';
+// @flow
+import {type Person} from './person';
+import {type Picture} from './picture';
 import {getWpFeaturedImage} from './media';
 import {bodyParser} from '../util/body-parser';
 
-// TODO: This needs a better model.
-export default class Article {
-  constructor(headline, articleBody, mainImage, associatedMedia, author, mainMedia) {
-    this.headline = headline;
-    this.articleBody = articleBody;
-    this.mainImage = mainImage;
-    this.associatedMedia = associatedMedia;
-    this.author = author;
-    this.mainMedia = mainMedia;
-    this.bodyParts = bodyParser(articleBody);
+export type BodyPart = {};
+
+export type Article = {|
+  headline: string;
+  standfirst: string;
+  // TODO: this will probably, at some stage be able to be video/audio/gallery etc
+  // It's also an Array because it's not unfathomable to think of having
+  // an audio and image mainMedia.
+  mainMedia: Array<Picture>;
+  articleBody: string;
+  associatedMedia: Array<Picture>;
+  author: Person;
+  bodyParts: Array<BodyPart>;
+|}
+
+function createArticle(data: Article) { return (data: Article); }
+
+export class ArticleFactory {
+  static fromWpApi(json): Article {
+    const articleBody = json.content;
+
+    const mainImage: Picture = getWpFeaturedImage(json.featured_image, json.attachments);
+
+    const author: Person = {
+      givenName: json.author.first_name,
+      familyName: json.author.last_name,
+      name: `${json.author.first_name} ${json.author.last_name}`,
+      image: json.author.avatar_URL,
+      sameAs: [{ wordpress: json.author.URL }]
+    };
+
+    const bodyParts = bodyParser(articleBody);
+    const standfirst = bodyParts.find(part => part.type === 'standfirst');
+
+    const article: Article = {
+      headline: json.title,
+      standfirst: standfirst,
+      mainMedia: [mainImage],
+      articleBody: articleBody,
+      associatedMedia: [mainImage],
+      author: author,
+      bodyParts: bodyParts
+    };
+
+    return article;
   }
-
-  static fromDrupalApi(json) {
-    return new Article(json.headline, json.articleBody, json.mainImage, json.associatedMedia);
-  }
-
-  static fromWpApi(json) {
-      const mainImage = getWpFeaturedImage(json.featured_image, json.attachments);
-      const author: Person = createPerson({
-        givenName: json.author.first_name,
-        familyName: json.author.last_name,
-        name: `${json.author.first_name} ${json.author.last_name}`,
-        image: json.author.avatar_URL,
-        sameAs: [{ wordpress: json.author.URL }]
-      });
-      return new Article(entities.decode(json.title), json.content, mainImage, [mainImage], author, [mainImage]);
-  }
-}
-
-function getMainMedia(associatedMedia) {
-  // This is to make sure we don't serve the main image and the main video
-  const video = associatedMedia.find(m => m.weighting === 'leading' && m.mediaType === 'video');
-  const image = associatedMedia.find(m => m.weighting === 'leading' && m.mediaType === 'image');
-  const audio = associatedMedia.find(m => m.weighting === 'leading' && m.mediaType === 'audio');
-  const gallery = associatedMedia.find(m => m.weighting === 'leading' && m.mediaType === 'gallery');
-
-  return [(gallery || video || image), audio];
 }
