+++ server/routes/index.js
@@ -8,7 +8,6 @@ r.get('/', index);
 r.get('/healthcheck', healthcheck);
 r.get('/explore', explore);
 r.get('/articles/:id', article);
-r.get('/artefacts/:id*', artefact);
 r.get('/performance-test.js', performanceTest);
 r.get('/explosion/:errorCode', explosion);
