+++ server/test/util/article.js
@@ -1,6 +1,6 @@
 import test from 'ava';
 import wpApiResp from '../mocks/wp-api.json';
-import {ArticleFactory} from '../../model/articleV2';
+import {ArticleFactory} from '../../model/article';
 
 test('articleV2', t => {
   const A = ArticleFactory.fromWpApi(wpApiResp);
